rewrite front UI use ng zorro

+++ angularapp/src/app/features/student/pages/students-table/students-table.component.ts
@@ -1,36 +1,47 @@
-import { Component, OnInit } from '@angular/core';
-import { MatDialog } from '@angular/material/dialog';
+import {
+  ChangeDetectionStrategy,
+  Component,
+  OnInit,
+  ViewContainerRef,
+} from '@angular/core';
 import { EditStudentDialogComponent } from '../../dialogs/edit-student-dialog/edit-student-dialog.component';
 import { Student } from 'src/app/core/models/responses/student';
-import { Observable } from 'rxjs';
+import {
+  Observable,
+  filter,
+  lastValueFrom,
+  map,
+  take,
+  takeWhile,
+  withLatestFrom,
+} from 'rxjs';
 import { Store } from '@ngrx/store';
 import {
+  selectErrors,
+  selectIsDeleteStudentLoading,
+  selectIsEditStudentLoading,
   selectIsStudentLoading,
   selectStudents,
 } from '../../store/selectors';
 import { getAllStudentsAction } from '../../store/actions/getAllStudents.action';
 import { deleteStudentAction } from '../../store/actions/deleteStudent.action';
+import { NzModalService } from 'ng-zorro-antd/modal';
 
 @Component({
   selector: 'app-students-table',
   templateUrl: './students-table.component.html',
   styleUrls: ['./students-table.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class StudentsTableComponent implements OnInit {
-  public displayedColumns = [
-    'firstName',
-    'lastName',
-    'middleName',
-    'age',
-    'group',
-    'actions',
-  ];
   students$!: Observable<Student[]>;
   isLoadingStudents$!: Observable<boolean>;
+  waitUntilEditLoading$!: Observable<boolean>;
 
   constructor(
-    public dialog: MatDialog,
-    private store: Store
+    private store: Store,
+    private modalService: NzModalService,
+    private viewContainerRef: ViewContainerRef
   ) {}
 
   ngOnInit(): void {
@@ -41,23 +52,54 @@ export class StudentsTableComponent implements OnInit {
   initializeValues(): void {
     this.students$ = this.store.select(selectStudents);
     this.isLoadingStudents$ = this.store.select(selectIsStudentLoading);
+
+    this.waitUntilEditLoading$ = this.store
+      .select(selectIsEditStudentLoading)
+      .pipe(
+        withLatestFrom(this.store.select(selectErrors)),
+        filter(value => !value[0]),
+        map(value => value[1] === null),
+        take(1)
+      );
   }
 
-  addNew() {
-    const dialogRef = this.dialog.open(EditStudentDialogComponent, {
-      disableClose: true,
-      data: null,
+  showAddNewStudentModal() {
+    this.modalService.create({
+      nzTitle: 'Добавить студента',
+      nzContent: EditStudentDialogComponent,
+      nzViewContainerRef: this.viewContainerRef,
+      nzData: null,
+      nzOkText: 'Сохранить',
+      nzOnOk: _ => {
+        return _.onSubmit(this.waitUntilEditLoading$);
+      },
     });
   }
 
-  startEdit(student: Student) {
-    const dialogRef = this.dialog.open(EditStudentDialogComponent, {
-      disableClose: true,
-      data: student,
+  showEditStudentModal(student: Student) {
+    this.modalService.create({
+      nzTitle: 'Редактировать студента',
+      nzContent: EditStudentDialogComponent,
+      nzData: student,
+      nzOkText: 'Сохранить',
+      nzOnOk: _ => {
+        return _.onSubmit(this.waitUntilEditLoading$);
+      },
     });
   }
 
-  deleteItem(id: string) {
-    this.store.dispatch(deleteStudentAction({ id }));
+  showConfirmDeleteStudentModal(id: string) {
+    this.modalService.confirm({
+      nzTitle: 'Вы уверены что хотите удалить студента?',
+      nzOnOk: () => {
+        this.store.dispatch(deleteStudentAction({ id }));
+
+        return lastValueFrom(
+          this.store
+            .select(selectIsDeleteStudentLoading)
+            .pipe(takeWhile(value => value))
+        );
+      },
+    });
   }
 }
